backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
langchain==0.3.7
langchain-huggingface==0.3.1
huggingface_hub==0.26.2
transformers==4.46.3
torch==2.5.1
langchain-pinecone==0.2.0
langchain-community==0.3.5
pinecone[grpc]==5.4.0
boto3==1.40.41
python-multipart==0.0.17
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
Pillow==11.0.0
mangum==0.19.0
